Move all types to a single file
Python's type system is barbaric compared to TypeScript's. In fact I
don't even currently have a checker installed, I'm just using it as
documentation. It's so barbaric that attempting to import types from
other files is sometimes treated as a circular import because there's
really no checking at all happening, at runtime or otherwise.

Moving all the types to their own file brute-forces my way out of that
problem and also gives me a handy declaration of all the types in the
system.
